requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
google-generativeai==0.3.0
python-dotenv==1.0.0
Werkzeug==2.3.7
